@@ -1,16 +1,16 @@
 name: flappy_search_bar
 description: A SearchBar widget automatizing most of your asynchronous searchs cases.
-version: 1.7.2
+version: 1.8.0
 homepage: https://github.com/smartnsoft/flappy_search_bar
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  async: ^2.2.0
-  flutter_staggered_grid_view: ^0.3.0
+  async: ^2.5.0
+  flutter_staggered_grid_view: ^0.4.0-nullsafety.3
 
 dev_dependencies:
   flutter_test:
