@@ -23,13 +23,13 @@ class SearchBarController<T> {
   final List<T> _list = [];
   final List<T> _filteredList = [];
   final List<T> _sortedList = [];
-  TextEditingController _searchQueryController;
-  String _lastSearchedText;
-  Future<List<T>> Function(String text) _lastSearchFunction;
-  _ControllerListener _controllerListener;
-  int Function(T a, T b) _lastSorting;
-  CancelableOperation _cancelableOperation;
-  int minimumChars;
+  late TextEditingController _searchQueryController;
+  String? _lastSearchedText;
+  Future<List<T>> Function(String? text)? _lastSearchFunction;
+  _ControllerListener? _controllerListener;
+  int Function(T a, T b)? _lastSorting;
+  CancelableOperation<List<T>>? _cancelableOperation;
+  late int minimumChars;
 
   void setTextController(TextEditingController _searchQueryController, minimunChars) {
     this._searchQueryController = _searchQueryController;
@@ -50,20 +50,21 @@ class SearchBarController<T> {
     _controllerListener.onListChanged(_list);
   }
   void _search(
-      String text, Future<List<T>> Function(String text) onSearch) async {
+      String? text, Future<List<T>> Function(String? text) onSearch) async {
     _controllerListener?.onLoading();
     try {
       if (_cancelableOperation != null &&
-          (!_cancelableOperation.isCompleted ||
-              !_cancelableOperation.isCanceled)) {
-        _cancelableOperation.cancel();
+          (!_cancelableOperation!.isCompleted ||
+              !_cancelableOperation!.isCanceled)) {
+        _cancelableOperation!.cancel();
       }
       _cancelableOperation = CancelableOperation.fromFuture(
         onSearch(text),
         onCancel: () => {},
       );
 
-      final List<T> items = await _cancelableOperation.value;
+
+      final List<T> items = await (_cancelableOperation!.value);
       _lastSearchFunction = onSearch;
       _lastSearchedText = text;
       _list.clear();
@@ -73,13 +74,13 @@ class SearchBarController<T> {
       _list.addAll(items);
       _controllerListener?.onListChanged(_list);
     } catch (error) {
-      _controllerListener?.onError(error);
+      _controllerListener?.onError(error as Error);
     }
   }
 
   void injectSearch(
-      String searchText, Future<List<T>> Function(String text) onSearch) {
-    if (searchText != null && searchText.length >= minimumChars) {
+      String searchText, Future<List<T>> Function(String? text) onSearch) {
+    if (searchText.length >= minimumChars) {
       _searchQueryController.text = searchText;
       _search(searchText, onSearch);
     }
@@ -87,7 +88,7 @@ class SearchBarController<T> {
 
   void replayLastSearch() {
     if (_lastSearchFunction != null && _lastSearchedText != null) {
-      _search(_lastSearchedText, _lastSearchFunction);
+      _search(_lastSearchedText, _lastSearchFunction!);
     }
   }
 
@@ -133,13 +134,13 @@ typedef ScaledTile IndexedScaledTileBuilder(int index);
 
 class SearchBar<T> extends StatefulWidget {
   /// Future returning searched items
-  final Future<List<T>> Function(String text) onSearch;
+  final Future<List<T>> Function(String? text) onSearch;
 
   /// List of items showed by default
   final List<T> suggestions;
 
   /// Callback returning the widget corresponding to a Suggestion item
-  final Widget Function(T item, int index) buildSuggestion;
+  final Widget Function(T item, int index)? buildSuggestion;
 
   /// Minimum number of chars required for a search
   final int minimumChars;
@@ -148,7 +149,7 @@ class SearchBar<T> extends StatefulWidget {
   final Widget Function(T item, int index) onItemFound;
 
   /// Callback returning the widget corresponding to an Error while searching
-  final Widget Function(Error error) onError;
+  final Widget Function(Error error)? onError;
 
   /// Cooldown between each call to avoid too many
   final Duration debounceDuration;
@@ -160,13 +161,13 @@ class SearchBar<T> extends StatefulWidget {
   final Widget emptyWidget;
 
   /// Widget to show by default
-  final Widget placeHolder;
+  final Widget? placeHolder;
 
   /// Widget showed on left of the search bar
   final Widget icon;
 
   /// Widget placed between the search bar and the results
-  final Widget header;
+  final Widget? header;
 
   /// Hint text of the search bar
   final String hintText;
@@ -184,10 +185,10 @@ class SearchBar<T> extends StatefulWidget {
   final Widget cancellationWidget;
 
   /// Callback when cancel button is triggered
-  final VoidCallback onCancelled;
+  final VoidCallback? onCancelled;
 
   /// Controller used to be able to sort, filter or replay the search
-  final SearchBarController searchBarController;
+  final SearchBarController? searchBarController;
 
   /// Enable to edit the style of the search bar
   final SearchBarStyle searchBarStyle;
@@ -200,7 +201,7 @@ class SearchBar<T> extends StatefulWidget {
 
   /// Called to get the tile at the specified index for the
   /// [SliverGridStaggeredTileLayout].
-  final IndexedScaledTileBuilder indexedScaledTileBuilder;
+  final IndexedScaledTileBuilder? indexedScaledTileBuilder;
 
   /// Set the scrollDirection
   final Axis scrollDirection;
@@ -221,9 +222,9 @@ class SearchBar<T> extends StatefulWidget {
   final EdgeInsetsGeometry listPadding;
 
   SearchBar({
-    Key key,
-    @required this.onSearch,
-    @required this.onItemFound,
+    Key? key,
+    required this.onSearch,
+    required this.onItemFound,
     this.searchBarController,
     this.minimumChars = 3,
     this.debounceDuration = const Duration(milliseconds: 500),
@@ -257,15 +258,15 @@ class SearchBar<T> extends StatefulWidget {
   _SearchBarState createState() => _SearchBarState<T>();
 }
 
-class _SearchBarState<T> extends State<SearchBar<T>>
-    with TickerProviderStateMixin, _ControllerListener<T> {
+class _SearchBarState<T> extends State<SearchBar<T?>>
+    with TickerProviderStateMixin, _ControllerListener<T?> {
   bool _loading = false;
-  Widget _error;
+  Widget? _error;
   final _searchQueryController = TextEditingController();
-  Timer _debounce;
+  Timer? _debounce;
   bool _animate = false;
-  List<T> _list = [];
-  SearchBarController searchBarController;
+  List<T?> _list = [];
+  late SearchBarController searchBarController;
 
   @override
   void initState() {
@@ -277,7 +278,7 @@ class _SearchBarState<T> extends State<SearchBar<T>>
   }
 
   @override
-  void onListChanged(List<T> items) {
+  void onListChanged(List<T?> items) {
     setState(() {
       _loading = false;
       _list = items;
@@ -302,17 +303,17 @@ class _SearchBarState<T> extends State<SearchBar<T>>
   void onError(Error error) {
     setState(() {
       _loading = false;
-      _error = widget.onError != null ? widget.onError(error) : Text("error");
+      _error = widget.onError != null ? widget.onError!(error) : Text("error");
     });
   }
 
   _onTextChanged(String newText) async {
     if (_debounce?.isActive ?? false) {
-      _debounce.cancel();
+      _debounce!.cancel();
     }
 
     _debounce = Timer(widget.debounceDuration, () async {
-      if (newText.length >= widget.minimumChars && widget.onSearch != null) {
+      if (newText.length >= widget.minimumChars) {
         searchBarController._search(newText, widget.onSearch);
       } else {
         setState(() {
@@ -327,7 +328,7 @@ class _SearchBarState<T> extends State<SearchBar<T>>
 
   void _cancel() {
     if (widget.onCancelled != null) {
-      widget.onCancelled();
+      widget.onCancelled!();
     }
 
     setState(() {
@@ -346,7 +347,7 @@ class _SearchBarState<T> extends State<SearchBar<T>>
   }
 
   Widget _buildListView(
-      List<T> items, Widget Function(T item, int index) builder) {
+      List<T?> items, Widget Function(T? item, int index) builder) {
     return Padding(
       padding: widget.listPadding,
       child: StaggeredGridView.countBuilder(
@@ -354,7 +355,7 @@ class _SearchBarState<T> extends State<SearchBar<T>>
         itemCount: items.length,
         shrinkWrap: widget.shrinkWrap,
         staggeredTileBuilder:
-            widget.indexedScaledTileBuilder ?? (int index) => ScaledTile.fit(1),
+            widget.indexedScaledTileBuilder ?? ((int index) => ScaledTile.fit(1)),
         scrollDirection: widget.scrollDirection,
         mainAxisSpacing: widget.mainAxisSpacing,
         crossAxisSpacing: widget.crossAxisSpacing,
@@ -366,7 +367,7 @@ class _SearchBarState<T> extends State<SearchBar<T>>
     );
   }
 
-  Widget _buildContent(BuildContext context) {
+  Widget? _buildContent(BuildContext context) {
     if (_error != null) {
       return _error;
     } else if (_loading) {
@@ -452,7 +453,7 @@ class _SearchBarState<T> extends State<SearchBar<T>>
           child: widget.header ?? Container(),
         ),
         Expanded(
-          child: _buildContent(context),
+          child: _buildContent(context)!,
         ),
       ],
     );
